Move to seaprate model file
