+++ .github/workflows/pr-python-checks.yml
@@ -48,7 +48,6 @@ jobs:
           -g python \
           -o /local/onyx_openapi_client \
           --package-name onyx_openapi_client \
-          --additional-properties=disallowAdditionalPropertiesIfNotPresent=false
             
     - name: Run MyPy
       run: |
